+++ src/audiovideocontroller/AudioVideoControllerState.ts
@@ -148,6 +148,9 @@ export default class AudioVideoControllerState {
   // that is provided to the peer connection, which will be ordered by priority.
   prioritizedSendVideoCodecCapabilities: VideoCodecCapability[] = [];
 
+  // Video codecs blocklisted to be used for sending due to encoding issues
+  videoSendCodecsBlocklisted: VideoCodecCapability[] = [];
+
   videosPaused: VideoStreamIdSet | null = null;
 
   videoDuplexMode: SdkStreamServiceType | null = null;