+++ src/connectionhealthpolicy/ConnectionHealthPolicyConfiguration.ts
@@ -1,6 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import VideoEncodingCpuConnectionHealthPolicy from '../connectionhealthpolicy/VideoEncodingCpuConnectionHealthPolicy';
+import VideoEncodingFramerateConnectionHealthPolicy from '../connectionhealthpolicy/VideoEncodingFramerateConnectionHealthPolicy';
+
 export default class ConnectionHealthPolicyConfiguration {
   minHealth: number = 0;
   maxHealth: number = 1;
@@ -39,4 +42,15 @@ export default class ConnectionHealthPolicyConfiguration {
    * as the microphone may sometimes cause a delay in sending audio packets during the initial stages of a connection.
    */
   sendingAudioFailureInitialWaitTimeMs = 3000;
+
+  /**
+   * Policies and parameters related to video encoding health montoring
+   */
+  videoEncodingHealthPolicies = [
+    VideoEncodingCpuConnectionHealthPolicy,
+    VideoEncodingFramerateConnectionHealthPolicy,
+  ];
+  consecutiveHighEncodeCpuThreshold = 10;
+  highEncodeCpuMsThreshold = 500;
+  consecutiveVideoEncodingFailureThreshold = 5;
 }